Hooking up ADC and trimpot.

+++ BirdsOfTheDamned/DamnedState/src/sc_types.h
@@ -18,7 +18,7 @@ typedef uint_fast16_t	sc_ushort;
 typedef int32_t			sc_integer;
 typedef uint32_t		sc_uinteger;
 
-typedef double			sc_real;
+typedef float			sc_real;
 
 typedef void*			sc_eventid;
